requirements.txt
PyYAML
argparse
ops
